yolox_ti_lite_tflite_int8/requirements.txt
# TODO: Update with exact module version
numpy==1.23.5
opencv_python
loguru
scikit-image
tqdm
torchvision
Pillow
thop
ninja
tabulate
tensorboard
plyfile
seaborn

# https://github.com/Megvii-BaseDetection/YOLOX/blob/618fd8c08b2bc5fac9ffbb19a3b7e039ea0d5b9a/requirements.txt#L16-L18
protobuf==3.20.3
onnx==1.16.0
onnx-simplifier==0.4.10
pycocotools>=2.0.2
onnxruntime>=0.14.1
# ONNX2TF
onnx-graphsurgeon==0.5.2
tensorflow==2.12.0
sng4onnx
onnx2tf==1.8.1
h5py==3.11.0
# UI
ipyfilechooser==0.6.0
ipywidgets==8.1.3
jupyterlab
